fix: move tests into relative packages, rename health to status
