+++ pkg/status/build.go
@@ -1,17 +1,8 @@
-package health
+package status
 
 import (
-	"encoding/json"
-	"net/http"
 	"runtime/debug"
 )
-
-const okValue = "ok"
-
-type Status struct {
-	Status    string     `json:"status"`
-	BuildInfo *BuildInfo `json:"buildInfo"`
-}
 
 type BuildInfo struct {
 	Version string `json:"version"`
@@ -40,21 +31,4 @@ func gitRevision(settings []debug.BuildSetting) string {
 	}
 
 	return "n/a"
-}
-
-func HandleAlive(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-
-	rr := Status{
-		Status: okValue,
-	}
-
-	if buildInfo := buildInfo(); buildInfo != nil {
-		rr.BuildInfo = buildInfo
-	}
-
-	json.NewEncoder(w).Encode(rr)
-
-	return
 }
